GEOMESA-1865 Refactor bin file computation for common use
